+++ src/core/energy_storage/electric_storage.jl
@@ -185,6 +185,8 @@ end
     model_degradation::Bool = false
     degradation::Dict = Dict()
     minimum_avg_soc_fraction::Float64 = 0.0
+    min_duration_hours::Real = 0.0 # Minimum amount of time storage can discharge at its rated power capacity
+    max_duration_hours::Real = 100000.0 # Maximum amount of time storage can discharge at its rated power capacity (ratio of ElectricStorage size_kwh to size_kw)
 ```
 """
 Base.@kwdef struct ElectricStorageDefaults
@@ -218,6 +220,8 @@ Base.@kwdef struct ElectricStorageDefaults
     model_degradation::Bool = false
     degradation::Dict = Dict()
     minimum_avg_soc_fraction::Float64 = 0.0
+    min_duration_hours::Real = 0.0
+    max_duration_hours::Real = 100000.0
 end
 
 
@@ -259,6 +263,8 @@ struct ElectricStorage <: AbstractElectricStorage
     model_degradation::Bool
     degradation::Degradation
     minimum_avg_soc_fraction::Float64
+    min_duration_hours::Real
+    max_duration_hours::Real
 
     function ElectricStorage(d::Dict, f::Financial)  
         s = ElectricStorageDefaults(;d...)
@@ -280,6 +286,10 @@ struct ElectricStorage <: AbstractElectricStorage
             end
             replace_cost_per_kwh = 0.0 # Always modeled using maintenance_cost_vector in degradation model.
             # replace_cost_per_kw is unchanged here.
+        end
+
+        if s.min_duration_hours > s.max_duration_hours
+            throw(@error("ElectricStorage min_duration_hours must be less than max_duration_hours."))
         end
 
         net_present_cost_per_kw = effective_cost(;
@@ -345,7 +355,9 @@ struct ElectricStorage <: AbstractElectricStorage
             net_present_cost_per_kwh,
             s.model_degradation,
             degr,
-            s.minimum_avg_soc_fraction
+            s.minimum_avg_soc_fraction,
+            s.min_duration_hours,
+            s.max_duration_hours
         )
     end
 end